test: Rename test resource folder
