+++ src/commonTest/resources/cases/auth/approle/create_custom_secret_id/with_options/expected_write.json
