+++ src/commonTest/resources/cases/auth/approle/create_custom_secret_id/without_options/expected_write.json
