+++ src/commonTest/resources/cases/auth/approle/generate_secret_id/with_options/expected_read.json
