+++ src/commonTest/resources/cases/auth/approle/generate_secret_id/with_options/given.json
