+++ src/commonTest/resources/cases/auth/approle/generate_secret_id/without_options/expected_read.json
