+++ src/commonTest/resources/cases/auth/approle/read_secret_id_accessor/without_options/expected_read.json
