+++ src/commonTest/resources/cases/auth/token/create_role/with_options/given.json
