+++ src/commonTest/resources/cases/auth/token/lookup_accessor/with_options/expected.json
