+++ src/commonTest/resources/cases/auth/token/lookup_accessor/without_options/expected.json
