+++ src/commonTest/resources/cases/auth/token/lookup_self/with_options/expected.json
