+++ src/commonTest/resources/cases/auth/token/lookup_self/without_options/expected.json
